+++ .gitignore
@@ -7,3 +7,6 @@
 target/
 /project/target/*
 *.csv
+derby.log
+metastore_db
+spark-warehouse
