+++ src/main/scala/com/data/InfectionData.scala
@@ -1,0 +1,5 @@
+package com.data
+
+object InfectionData {
+
+}
