+++ src/main/scala/com/data/MortalityData.scala
@@ -1,0 +1,5 @@
+package com.data
+
+object MortalityData {
+
+}
